+++ src/privateer/restraints.py
@@ -1,6 +1,6 @@
-from privateer.pyprivateer import check_monlib_access
-from privateer.pyprivateer import CarbohydrateDictionary
-from privateer.pyprivateer import CarbohydrateLibrary
+from privateer.libprivateer import check_monlib_access
+from privateer.libprivateer import CarbohydrateDictionary
+from privateer.libprivateer import CarbohydrateLibrary
 import os
 
 def minimise_from_smiles ( smiles_string = "", n_conformers=50, filename="privateer-minimised.pdb" ) :
